+++ CHANGELOG.md
@@ -4,15 +4,16 @@
 - 全文検索（投稿や回答があったときにデータベースをアップデートする仕組み）
 - qa, qa-all で ol の自動番号の代わりにテーブルの id にしたらどうか？
 - 質問も markdown?
-- app.melt で systemctl stop qa できてないか？
-  restart.sh だとリスタートできる。
+- app.melt で systemctl stop qa できてない。
+　systemctl ではなく、restart.sh だとリスタートできる。
 - (reset) で毎回、クラッシュ。lein repl だとクラッシュは減る。
   duct じゃなく、VScode の REPL がダメか？
 
 
 ## 1.6.0-SNAPSHOT
 ### Changed
-- - q/a のテキストエリアの高さを 2 倍、200px に。
+- q/a のテキストエリアの高さを 2 倍、200px
+- /since を hkimura オンリーに
 
 ## 1.5.3 - 2022-07-02
 ### Added